--- spongebob.py
@@ -1,4 +1,0 @@
-from samples import helix
-
-if __name__ == '__main__':
-    helix.first_500_streams()
