+++ twitch/__init__.py
@@ -1,4 +1,4 @@
 from .client import TwitchClient  # noqa
-from .helix import TwitchHelix  # noqa
+from .helix.api import TwitchHelix  # noqa
 
-__version__ = '0.5.0'
+__version__ = '0.5.1'
